Refactor the factories into a module with sub-modules for each factory.  moved drape_electrodes_on_topography call into survey factory.
