+++ geoapps/drivers/components/factories/__init__.py
@@ -1,0 +1,9 @@
+#  Copyright (c) 2021 Mira Geoscience Ltd.
+#
+#  This file is part of geoapps.
+#
+#  geoapps is distributed under the terms and conditions of the MIT License
+#  (see LICENSE file at the root of this source code package).
+
+from .simulation_factory import SimulationFactory
+from .survey_factory import SurveyFactory
